Move post-lock into post detail and rename post-edit for splitting later
