+++ legacy/app/main/data/post-detail/post-lock.directive.js
