+++ legacy/app/main/data/post-detail/post-lock.service.js
