+++ legacy/app/main/data/split-into-post-edit-and-create/post-create.controller.js
