+++ legacy/app/main/data/split-into-post-edit-and-create/post-data-editor.directive.js
