+++ legacy/app/main/data/split-into-post-edit-and-create/post-tabs.directive.js
