+++ legacy/app/main/data/split-into-post-edit-and-create/post-toolbox.html
